avoid developer builds to generate artifacts
