+++ scripts/build.js
@@ -51,12 +51,19 @@ const disabledTasks = getDisabledTasks(process, package.disabledTasks);
 const firstTasks = getNextTasks(null, disabledTasks);
 
 // Start executing tasks, executeTasks will call itself recursively until all tasks are done
-executeTasks(firstTasks).then(() => {
-  if (hasFailures) {
-    process.exitCode = 1;
-  }
-  logEndBuild(packageName, !hasFailures, buildStartTime);
-}).then(() => executeTask());
+executeTasks(firstTasks)
+  .then(() => {
+    if (hasFailures) {
+      process.exitCode = 1;
+    }
+    logEndBuild(packageName, !hasFailures, buildStartTime);
+  })
+  .then(() => {
+    if (process.env['CI_BUILD']) {
+      const { generateSizeData } = require('./tasks/size-audit');
+      generateSizeData();
+    }
+  });
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////
 //// Build helper functions
@@ -167,45 +174,4 @@ function getDisabledTasks(process, defaultDisabled = []) {
   }
 
   return defaultDisabled;
-}
-/////////////////////////////////////////////////////////////////////////////////////////////////////
-//// Artifact helper functions
-////////////////////////////////////////////////////////////////////////////////////////////////////
-/**
-   * Gets the size of the "default" locale bundle file
-   */
-function executeTask() {
-  const projectName = path.basename(process.cwd());
-  const sizeFilePath = path.join(process.cwd(), 'dist', `${projectName}.json`);
-  if (fs.existsSync(path.dirname(sizeFilePath))) {
-    const result = {
-      chunks: analyzeChunks(),
-    };
-    const fileContents = JSON.stringify(result);
-    fs.writeFileSync(sizeFilePath, fileContents);
-  }
-}
-
-function analyzeChunks() {
-  const distPath = path.join(process.cwd(), 'dist');
-  const result = {};
-  for (const filePath of fs.readdirSync(distPath)) {
-    const fileName = path.basename(filePath);
-    const extension = path.extname(fileName);
-    if (extension.toLowerCase() === '.js') {
-      result[fileName] = getFileSize(path.join(distPath, fileName));
-    }
-    return result;
-  }
-}
-
-function getFileSize(filePath) {
-  try {
-    const stats = fs.statSync(filePath);
-    return stats.size;
-  }
-  catch (e) {
-    console.log(`Unable to get size of file "${filePath}"`);
-    return -1;
-  }
 }
